+++ src/core/index.ts
@@ -1,3 +1,4 @@
+export * from "./burst";
 export * from "./constants";
 export * from "./fragments";
 export * from "./lanes";